+++ CHANGELOG.md
@@ -11,6 +11,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add support for dispatching jobs to the queue after the current response has
   been sent.
+- Allow the block factory to override text when generating blocks.
+
+## v1.1.3 - 2024-08-14
+
+### Added
+
+- Added a `with_image()` helper to mocked HTTP responses.
+- Added a `is_blob()` and `is_file()` helper to `Mantle\Http_Client\Response`.
+- Added `with_real_thumbnail()` method to post factory for creating posts with
+  real underlying thumbnail files.
+
+### Changed
+
+- Added support for faking specific HTTP requests by method.
+- Added helper for fluently building HTTP sequence responses.
+
+### Fixed
+
+- Fixed an issue with taxonomy registration not returning an array.
 
 ## v1.1.2 - 2024-06-20